# [Enhancement] Optimize the performance of BitmapValue::add()