Replace _is_shared with use_count of std::shared_ptr, It has no impact on performance and can make the code more concise.

Move BitmapValue::add() to bitmap_value.h to improve the performance of optimize the performance of high-frequency calls function such as bitmap_agg(bigint)

lo_orderkey is bigint type

```
 select bitmap_count(bitmap_agg(lo_orderkey)) from lineorder_bigint;
```

TotalTime: 5.22s -> 4.76s
AggFuncComputeTime: 5.066s -> 4.594s

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5