@trueeyu 2023-11-15
https://github.com/Mergifyio backport branch-3.2